requirements.txt
pynetbox
python_powerdns
